Latest v1.9.0 (#135)
Provide dummy CPM-based packaging impl for ut

Refactor common compiler settings

Use rt_common_settings for riptide_bench

Avoid inheriting ut compiler warnings

Rearrange how expected awoken threads are awaited

Add quantile and nanquantile to groupby operations

Pin tbb to 2021.6

Groupby min/max/nanmin/nanmax fix

Enforce binutils v2.36

+++ src/SDSFile.cpp
@@ -241,7 +241,7 @@ size_t ZSTD_decompress_stackmode(void * dst, size_t dstCapacity, const void * sr
     size_t regenSize;
     ZSTD_DCtx * const dctx = ZSTD_createDCtx();
     if (dctx == NULL)
-        return -64; // ERROR(memory_allocation);
+        return static_cast<size_t>(-64); // ERROR(memory_allocation);
     regenSize = ZSTD_decompressDCtx(dctx, dst, dstCapacity, src, srcSize);
     ZSTD_freeDCtx(dctx);
     return regenSize;
@@ -317,7 +317,7 @@ static size_t DecompressDataPartial(int32_t core, int32_t compMode, void * dst, 
 
         // return regenSize;
     }
-    return -1;
+    return static_cast<size_t>(-1);
 }
 
 //------------------------------------------------------------
